+++ wasmvm_test.go
